pyproject.toml
[project]
name = "clip"
version = "0.1.0"
description = "CLIP training from scratch — MLE learning project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "open-clip-torch>=2.26.1",
    "tensorboard>=2.18",
    "umap-learn>=0.5",
    "matplotlib>=3.9",
    "aiohttp>=3.9",
    "datasets>=4.8.3",
    "jupyter>=1.1.1",
    "ipykernel>=7.2.0",
]

[tool.uv]
find-links = ["https://download.pytorch.org/whl/cu124"]
